requirements.txt
openai>=1.3.0
python-dotenv>=1.0.0
openai-agents>=0.0.16
chromadb>=0.4.22
fastapi>=0.104.1
uvicorn>=0.24.0
zipfile36>=0.1.3
aiohttp>=3.9.0
asyncio
